Main code update 6/15
This version adds functions that allow user input for budget, reach, and convergence threshold (although the last still needs to get plugged in to the rest of the system) from command line, so it doesn't have to be recompiled every time.
I also wrote a couple functions that, given a list of kets to be removed and kets to be added, updates the reference bases opBasis and matrixBasis and updated Htrunc. Whenever we have new schemes for exploring and rearranging the Hilbert space, we can use these functions to update the basis.